Remove duplicate test cases, and move test cases for infrequently used features to the full-test pipeline.
