+++ test/registered/npu/basic_function/speculative_inference/test_npu_speculative_multi_npu.py
@@ -16,8 +16,7 @@ from sglang.test.test_utils import (
     popen_launch_server,
 )
 
-register_npu_ci(est_time=400, suite="base-b-test-4-npu-a3")
-register_npu_ci(est_time=400, suite="nightly-4-npu-a3", nightly=True)
+register_npu_ci(est_time=400, suite="full-4-npu-a3", nightly=True)
 
 _ASCEND_BACKEND = "ascend"
